+++ drand_core/CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- Change RFC 3339 datetime parsing to allow space and date
+
+### Fix
+
 - Fix clippy warning for wasm32-unknown-unknown target
 
 ## [0.0.15] - 2024-02-29